send email and create laravel event for this

+++ learn-laravel/app/Http/Controllers/TodoController.php
@@ -4,16 +4,38 @@ namespace App\Http\Controllers;
 
 use App\Todo;
 use App\Services\Twitter;  
+use App\Events\TodoCreated;
 
 use Illuminate\Http\Request;
 
 class TodoController extends Controller
 {
+    public function __construct() {
+
+        $this->middleware('auth'); // ->except(['show']); - не авторизированным пользователям можно просматривать задачи
+
+    }
+
     public function index() {
 
-        $tasks = Todo::all();
+        // auth()->id(); // id user
+        // auth()->user(); // get User
+        // auth()->check(); // check user is login - boolean
+        // auth()->guest(); // check user is guest
 
-        return view('todos.index', compact('tasks'));
+        // $tasks = Todo::where('owner_id', auth()->id())->get();
+        // or
+
+        // $todos = auth()->user()->todos;
+
+        // return view('todos.index', compact('todos'));
+
+        // or
+
+        return view('todos.index', [
+            // get all todos for the auth'd user 
+            'todos' => auth()->user()->todos
+        ]);
 
     }
 
@@ -25,18 +47,39 @@ class TodoController extends Controller
 
     public function store() {
 
-        $validated = request()->validate([
-            'todo_title' => ['required', 'min:3', 'max:255'],
-            'todo_description' => ['required', 'min:3']
-        ]);
+        $attributes = $this->validateTodo();
 
-        Todo::create($validated);
+        $attributes['owner_id'] = auth()->id();
+
+        $todo = Todo::create($attributes);
+
+        // event(new TodoCreated($todo));
 
         return redirect('/todos');
 
     }
 
     public function show(Todo $todo) {
+
+        // if($todo->owner_id !== auth()->id()) {
+        //     abort(403); // not autorized for this page
+        // }
+        // or
+        // abort_if($todo->owner_id !== auth()->id(), 403);
+        // or
+        // abort_unless(auth()->user()->owns($todo), 403);
+        // OR
+        // if(\Gate::denies('update', $todo)) {
+        //     abort(403);
+        // }
+        // OR
+        // abort_if(\Gate::denies('update', $todo), 403);
+        // OR
+        // abort_unless(\Gate::allows('update', $todo), 403);
+
+        // или используем Policy
+
+        $this->authorize('view', $todo);
 
         return view('todos.show', compact('todo'));
     
@@ -50,16 +93,27 @@ class TodoController extends Controller
 
     public function update(Todo $todo) {
 
-        $todo->update(request(['todo_title', 'todo_description']));
+        $todo->update(request($this->validateTodo()));
 
         return redirect('/todos');
     }
 
     public function destroy(Todo $todo) {
+
+        $this->authorize('destroy', $todo);
         
         $todo->delete();
 
         return redirect('/todos');
+    }
+
+    public function validateTodo() {
+
+        return request()->validate([
+            'todo_title' => ['required', 'min:3', 'max:255'],
+            'todo_description' => ['required', 'min:3']
+        ]);
+
     }
 
 }
